+++ Runtime/Scripts/LsLeanCameraDollySmooth.cs
@@ -6,7 +6,7 @@ namespace Eos.Ux.Lean
     [ExecuteInEditMode]
     public class LsLeanCameraDollySmooth : LsLeanCameraDolly
     {
-        public float _dampening = 3.0f;
+        public float _dampening = 6.0f;
 
         private float _currentDolly;